requirements.txt

flask==2.0.1
werkzeug==2.0.3
redis==4.5.4
python-dotenv==0.19.0
requests==2.26.0

# (venv) PS J:\Meu DrIve\ProjetosPython\LoterIas\Estrategias\MegaSena\PalpitesMS> pip list
# Package            Version
# ------------------ ---------
# certifi            2025.1.31
# charset-normalizer 2.0.12
# click              8.1.8
# colorama           0.4.6
# Flask              2.0.1
# idna               3.10
# itsdangerous       2.2.0
# Jinja2             3.1.5
# MarkupSafe         3.0.2
# pip                25.0.1
# python-dotenv      0.19.0
# redis              4.5.4
# requests           2.26.0
# urllib3            1.26.20
# Werkzeug           2.0.3




